Moved antibody display_title into own test file and added kind of redundant _update tests
